spaceAPI/pyproject.toml
[tool.poetry]
name = "spaceapi"
version = "0.1.0"
description = ""
authors = ["Juan Esteban Escobar Restrepo <138255902+YugenDev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
requests = "^2.32.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
